Fixed tournament method
Previously there was a problem in the tournament method: sometimes there was an odd numbers of candidates  and that caused issues in the tournament because one candidate always passed directly in that cases. Now if the number of candidates is odd, there will be an odd group, where only one candidate will win. For instance if the initial group of candidates is 5, there will be two groups: one with 2 candidates and another with 3. This way it is not possible that a candidate passes without compiting,