General changes
In evolutionary algorithm has been added a system to register all the individuals which are generated during the process so that it is easier to analyse the results. Furthermore, the fitness function has been changed in order to reduce the scale's difference between standard deviation and mean of the vertx elements

In individual class has been added a new attribute to identify individuals, because the directory as identifier difficulted the analysis. 

Finally, generational change has now a limit of individuals to generate with mutation and with crossover.