General changes
The instruction System.gc() has been deleted since may produce some errors. A new property has been added to individual, the average volume of the segmentation so that it is possible to distinguish solutions that may be under segmented. The fitness function has also been changed so that the volume is taken into account.