General Changes
The evolutionary algorithm class has been changed  because a new method has been added to explore the possibilies of centroid positions. Moreover, Individuo class has also changed to add more properties. It is important to highlight the change in SphereSegAdapted, where the number of steps in the optimization has been modified to put it like it was initially.